Remove unnecesary login files and get oauth clientId correctly
