+++ .idea/modules.xml
@@ -3,7 +3,6 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/firebasetraining.iml" filepath="$PROJECT_DIR$/.idea/firebasetraining.iml" />
       <module fileurl="file://$PROJECT_DIR$/firebasetraining.iml" filepath="$PROJECT_DIR$/firebasetraining.iml" />
     </modules>
